follow DHH way of application controller
