+++ app/controllers/concerns/protect_from_forgery.rb
@@ -1,0 +1,7 @@
+module ProtectFromForgery
+  extend ActiveSupport::Concern
+
+  included do
+    protect_from_forgery with: :exception, prepend: true
+  end
+end
