+++ app/helpers/masquerading_helper.rb
@@ -1,0 +1,5 @@
+module MasqueradingHelper
+  def masquerading?
+    session[:admin_id].present?
+  end
+end
